perf: reduce int size (#87) [skip ci]

+++ contracts/mocks/DCAHub/DCAHubParameters.sol
@@ -60,8 +60,8 @@ contract DCAHubParametersMock is DCAHubParameters {
     address _tokenA,
     address _tokenB,
     uint32 _swapInterval,
-    uint256 _amountToSwapAToB,
-    uint256 _amountToSwapBToA
+    uint224 _amountToSwapAToB,
+    uint224 _amountToSwapBToA
   ) external {
     swapData[_tokenA][_tokenB][intervalToMask(_swapInterval)].nextAmountToSwapAToB = _amountToSwapAToB;
     swapData[_tokenA][_tokenB][intervalToMask(_swapInterval)].nextAmountToSwapBToA = _amountToSwapBToA;
